placeless-view/requirements.txt
streamlit
pandas
altair
numpy
mysql-connector
